+++ src/rpc/net.cpp
@@ -120,6 +120,13 @@ static UniValue getpeerinfo(const Config &config,
                      "The " + UNIX_EPOCH_TIME + " of the last send"},
                     {RPCResult::Type::NUM_TIME, "lastrecv",
                      "The " + UNIX_EPOCH_TIME + " of the last receive"},
+                    {RPCResult::Type::NUM_TIME, "last_transaction",
+                     "The " + UNIX_EPOCH_TIME +
+                         " of the last valid transaction received from this "
+                         "peer"},
+                    {RPCResult::Type::NUM_TIME, "last_block",
+                     "The " + UNIX_EPOCH_TIME +
+                         " of the last block received from this peer"},
                     {RPCResult::Type::NUM, "bytessent", "The total bytes sent"},
                     {RPCResult::Type::NUM, "bytesrecv",
                      "The total bytes received"},
@@ -222,6 +229,8 @@ static UniValue getpeerinfo(const Config &config,
         obj.pushKV("relaytxes", stats.fRelayTxes);
         obj.pushKV("lastsend", stats.nLastSend);
         obj.pushKV("lastrecv", stats.nLastRecv);
+        obj.pushKV("last_transaction", stats.nLastTXTime);
+        obj.pushKV("last_block", stats.nLastBlockTime);
         obj.pushKV("bytessent", stats.nSendBytes);
         obj.pushKV("bytesrecv", stats.nRecvBytes);
         obj.pushKV("conntime", stats.nTimeConnected);